Exception is thrown with invalid API key.
The library throws a special exception if the constructor is supplied
with and invalid API key upon initialisation. The HTTP requests for the
user's data have been factored out into a separate method which checks
the HTTP status code of the response and raises an exception if it is
equal to 401 (Unauthorized).