Fix edit_device method to keep nickname
Editing a device using the edit_device() method without a nickname given would result in the nickname being deleted from the device. This is due to the default nickname=None value. This also then causes a 'tuple index out of range' error if the device list is called upon using pb.devices.