Add new method to get device by nickname
Currently to retrieve a specific device, the user needs to know the index of the `Device` in the device list. This requires either to print devices first using `pb.devices` or to retrieve all devices and use a loop.

As the pushbullet.py package retrieves all devices on initial connection, it is easier to provide a helper method for the end user.